# apps/backend/src/shared/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Mandarin Learning Platform API
  version: 1.0.0
  description: |
    Production backend API for Mandarin vocabulary learning and AI-powered
    conversation generation. All endpoints are versioned under /api/v1.

    This reference documents the full NestJS 11 route surface: 54 route paths
    across the 15 backend modules (audio, auth, characters, chengyu,
    foundations, grammar, health, mnemonics, phonetic-clusters, progression,
    quiz, radicals, readers, review, words — pinyin lives under foundations).

    Authentication model (per route):
      - Public — no auth required (static reference data, health, auth
        register/login/refresh/logout).
      - Required auth — `Authorization: Bearer <access_token>`; user-scoped
        state or AI/vendor-cost generation (`RequireAuthGuard`).
      - Optional auth — bearer token is best-effort (`OptionalAuthGuard`): an
        absent or invalid token proceeds as a guest (`req.userId` undefined),
        never 401. Guest responses are session-local/empty shapes.

    All 4xx/5xx errors use the `{ code, message, requestId }` envelope.
  contact:
    name: API Support
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://mandarin-vite-react-ts-production.up.railway.app/api
    description: Production server (Railway)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token for authenticated requests
    cookieAuth:
      type: apiKey
      in: cookie
      name: refreshToken
      description: HTTP-only refresh token cookie
  schemas:
    ErrorEnvelope:
      type: object
      description: |
        Standard error envelope — the global 24-3 `AppExceptionFilter`
        serializes every thrown `HttpException` into this shape.
      properties:
        code:
          type: string
          description: Machine-readable error code
          example: VALIDATION_ERROR
        message:
          type: string
          description: Human-readable error message
        requestId:
          type: string
          description: Request ID (echoes the `X-Request-Id` response header)
      required:
        - code
        - message
        - requestId
...
paths:
...
  /v1/review/items:
    get:
      summary: Get review items
      tags:
        - Review
      security:
        - bearerAuth: []
      description: |
        Fetches the authenticated user's review items. Required auth —
        user-scoped SRS reads.
      parameters:
        - name: source
          in: query
          required: false
          schema:
            type: string
          description: Source filter — due | recent | all (default due)
        - name: type
          in: query
          required: false
          schema:
            type: string
          description: Optional item-type prefix filter
        - name: limit
          in: query
          required: false
          schema:
            type: integer
          description: Result limit (default 20)
      responses:
        "200":
          description: Review items
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        "401":
          description: Missing or invalid JWT (AUTH_REQUIRED)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Failed to fetch review items (LOAD_FAILED)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /v1/review/result:
    post:
      summary: Record a review rating
      tags:
        - Review
      security:
        - bearerAuth: []
      description: |
        Records a rating (again/good/easy) for a review item and computes the
        next review date. Required auth — writes user-scoped SRS state.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - itemType
                - itemId
                - rating
              properties:
                itemType:
                  type: string
                  description: e.g. radical, character, word
                itemId:
                  type: string
                rating:
                  type: string
                  enum:
                    - again
                    - good
                    - easy
      responses:
        "200":
          description: Next review date / updated item result
          content:
            application/json:
              schema:
                type: object
        "400":
          description: Invalid itemType or rating (MISSING_FIELDS)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Missing or invalid JWT (AUTH_REQUIRED)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Failed to record rating (UPDATE_FAILED)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /v1/review/due-count:
    get:
      summary: Get due review count
      tags:
        - Review
      security:
        - bearerAuth: []
      description: Counts the authenticated user's due SRS rows. Required auth.
      parameters:
        - name: type
          in: query
          required: false
          schema:
            type: string
          description: Optional item-type prefix filter
      responses:
        "200":
          description: Due count
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
        "401":
          description: Missing or invalid JWT (AUTH_REQUIRED)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Failed to fetch due count (LOAD_FAILED)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # ══════════════════════════════════════════════════════════════════
  # Radicals
  # ══════════════════════════════════════════════════════════════════
  /v1/radicals:
    get:
      summary: List all radicals
      tags:
        - Radicals
      description: Returns all radicals from the Radical reference table. Public.
      responses:
        "200":
          description: All radicals
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        "500":
          description: Failed to load radicals (LOAD_ERROR)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /v1/radicals/{radicalId}:
    get:
      summary: Get a radical by ID
      tags:
        - Radicals
      description: |
        Returns a single radical by business-key ID (e.g. rad_0001). Unknown
        IDs return `200 null` — a byte-for-byte mirror of the Express
        `res.json(null)`. Public.
      parameters:
        - name: radicalId
          in: path
          required: true
          schema:
            type: string
          description: Radical business key (e.g. rad_0001)
      responses:
        "200":
          description: Radical object, or literal null for an unknown ID
          content:
            application/json:
              schema:
                type: object
                nullable: true
        "404":
          description: Load failure (NOT_FOUND)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /v1/radicals/character/{glyph}:
    get:
      summary: Get radicals composing a character
      tags:
        - Radicals
      description: Returns radicals that compose the given character glyph. Public.
      parameters:
        - name: glyph
          in: path
          required: true
          schema:
            type: string
          description: Single Chinese character glyph
      responses:
        "200":
          description: Radicals composing the character
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        "500":
          description: Failed to load radicals for character (LOAD_ERROR)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /v1/radicals/{radicalId}/characters:
    get:
      summary: Get characters associated with a radical
      tags:
        - Radicals
      description: Returns characters associated with the given radical. Public.
      parameters:
        - name: radicalId
          in: path
          required: true
          schema:
            type: string
          description: Radical business key (e.g. rad_0001)
      responses:
        "200":
          description: Characters for the radical
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        "404":
          description: Radical not found (NOT_FOUND)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Failed to load radical characters (LOAD_ERROR)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # ══════════════════════════════════════════════════════════════════
  # Foundations
  # ══════════════════════════════════════════════════════════════════
...